Added Key Features and API use
Changed it so there is a login page, this uses data from the JSON file, currently it is stored very basically (no hash or encryption). this is a good start and we can develop this in the future. Additionally, i changed it so the dashboard reflects the current users stocks, this is also saved under their account info in the JSON. Finally, I added an overview of news affecting their stocks, this can be viewed on the dashboard page. The news was done using an API which returns news related to the stock of the user.